Damage implemented

+++ Laser Defender/Assets/Scripts/DamageDealer.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class DamageDealer : MonoBehaviour
 {
-    [SerializeField] private int damage = 100;
+    [SerializeField] private float damage = 100f;
 
-    public int GetDamage()
+    public float GetDamage()
     {
         return damage;
     }
